Move stuff to utils.py
