+++ src/oemof/tabular/datapackage/utils.py
@@ -1,8 +1,35 @@
 import dataclasses
+import itertools
+import warnings
 
 import addict
+import pandas as pd
 
 from oemof.tabular.facades import TYPEMAP
+
+
+class Data:
+    """
+    Implements the hierarchical data as in the filetree.
+
+    data[][][] = resource
+    """
+
+    def __init__(self):
+        self.data = addict.Dict()
+
+    def set_resource(self, location, resource):
+        d = nest_dict_from_lst(location, resource)
+        self.data.update(d)
+
+    def get_resource(self, location):
+        result = self.data
+        for item in location:
+            result = result[item]
+        return result
+
+    def to_tuple_dict(self):
+        return tuple_keys(self.data)
 
 
 def get_facade_fields():
@@ -14,6 +41,28 @@ def get_facade_fields():
     }
 
     return facade_fields
+
+
+def populate_df(df, **kwargs):
+    undefined_keys = [key for key in kwargs if key not in df.columns]
+
+    if any(undefined_keys):
+        warnings.warn(f"There are undefined keys: {undefined_keys}")
+
+    # perform cartesian product if lists are given
+    given_values = {
+        k: v if isinstance(v, list) else [v] for k, v in kwargs.items()
+    }
+
+    cartesian_product = pd.DataFrame(
+        itertools.product(*given_values.values()), columns=given_values.keys()
+    )
+
+    # populate dataframe
+    for col in cartesian_product.columns:
+        df[col] = cartesian_product[col]
+
+    return df
 
 
 def tuple_keys(nested_dict):
